% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ssurgo_to_daycent.R
\name{convert_ssurgo_to_daycent}
\alias{convert_ssurgo_to_daycent}
\title{Convert SSURGO data to DayCent format}
\usage{
convert_ssurgo_to_daycent(latitude, longitude, site, templatesite)
}
\arguments{
\item{latitude}{numeric. Latitude of the site in decimal degrees.}

\item{longitude}{numeric. Longitude of the site in decimal degrees.}

\item{site}{character. Name of the site. Must correspond to an existing folder name.}

\item{templatesite}{character. Soils.in file template name.
Must correspond to an existing soils.in file to be used as a template.}
}
\value{
The function exports a soils.in file for the site of interest that is DayCent-ready.
}
\description{
This function converts SSURGO data into a DayCent-ready file format for specific locations.
SSURGO data is interpolated to fit the soil depth intervals for DayCent when necessary.
}
\details{
This function leverages the get_soil_properties function to obtain SSURGO data that is then converted into
a DayCent-ready file format for specific locations. Downloaded SSURGO data might not have all the soil layers necessary
for a DayCent run; this function interpolates the data for soil layers using a simple weighted approach based on
the overlap between the SSURGO and DayCent soil layer thickness.
The templatesite name should match the file name in the site folder where DayCent is being run.
The site name must correpond to the folder where the exported file will be saved.
The returned soils.in file follows the template format, and thus includes 13 variables:
\itemize{
\item upper_depth_cm, from soils.in template file.
\item lower_depth_cm, from soils.in template file.
\item dthirdbar_r, interpolated from SSURGO data.
\item wthirdbar_r, interpolated from SSURGO data.
\item wfifteenbar_r, interpolated from SSURGO data.
\item evap_coeff, from soils.in template file.
\item fract_roots, from soils.in template file.
\item sandtotal_r, interpolated from SSURGO data.
\item claytotal_r, interpolated from SSURGO data.
\item om_r, interpolated from SSURGO data.
\item min_fract, from soils.in template file.
\item ksat_r, interpolated from SSURGO data.
\item ph1to1h2o_r, interpolated from SSURGO data.
}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
latitude <- 45.7
longitude <- -118.6
site <- "test_site" ## this should be an existing folder within the sites folder for a DayCent run
templatesite <- "test_soils" ## this should be an existing soils.in to be used as a template
soils <- convert_ssurgo_to_daycent(latitude, longitude, site, templatesite) ## the file will be saved in the specified site folder.
\dontshow{\}) # examplesIf}
}
