% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overWriteSoil.R
\name{format_soil_block}
\alias{format_soil_block}
\title{Format soil layer parameters for a DayCent site file}
\usage{
format_soil_block(soil_df)
}
\arguments{
\item{soil_df}{Data frame produced by \code{\link{read_soil_in}}.}
}
\value{
Character vector of formatted DayCent soil parameter lines.
}
\description{
Converts a soil parameter data frame into DayCent-formatted text
lines suitable for insertion into a site (.100) file.
}
\examples{
\dontrun{
soil_df <- read_soil_in("soil.in")
block <- format_soil_block(soil_df)
cat(block, sep = "\n")
}

}
