% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lis_standard.R
\name{plot_lis_standard}
\alias{plot_lis_standard}
\title{Standard plots for lis output: SOC stock, aboveground biomass, and harvested grain C.}
\usage{
plot_lis_standard(lis_df, site = "", run = "", agb_col = "agcprd")
}
\arguments{
\item{lis_df}{data.frame. From build_lis_from_bin()$all (or similar).}

\item{site}{character. Used for plot titles.}

\item{run}{character. Used for plot titles.}

\item{agb_col}{character. Column name for aboveground biomass. Default is "agcprd".}
}
\value{
list of ggplot objects: p_soc, p_agb, p_cgrain.
}
\description{
Standard plots for lis output: SOC stock, aboveground biomass, and harvested grain C.
}
