% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicLisCheck.R
\name{plot_lis_standard}
\alias{plot_lis_standard}
\title{Standard plots for lis output: SOC stock + aboveground biomass}
\usage{
plot_lis_standard(lis_df, site = "", run = "", agb_col = "agcprd")
}
\arguments{
\item{lis_df}{data.frame from build_lis_from_bin()$all (or similar)}

\item{site}{character for titles}

\item{run}{character for titles}

\item{agb_col}{character. Column name for aboveground biomass (tries to guess if NULL)}
}
\value{
list of ggplot objects: p_soc, p_agb
}
\description{
Standard plots for lis output: SOC stock + aboveground biomass
}
